Fixed segmenting, data splitting and performed refactoring
This commit has the following changes:

(1) Segmenting the images now crops the image based on the bounding
box and doesn't save the entire image

(2) The saving of the images is now more intelligent given (1),
specifically to suit Keras ImageDataGenerator.  The images after
running them through Mask RCNN is that there are now clear
train and mask subdirectories and inside are subdirectories
containing each class

(3) Splitting the data into training and validation is also more
intelligent.  Specifically there will now be four subdirectories,
two for each of the segmented images and masks (train + validation)

(4) Refactored the data splitting code so that it's object oriented